+++ qiskit/providers/fake_provider/backends/__init__.py
@@ -62,7 +62,6 @@ from .valencia import FakeValenciaV2
 from .vigo import FakeVigoV2
 from .washington import FakeWashingtonV2
 from .yorktown import FakeYorktownV2
-from .generic import FakeGeneric
 
 # BackendV1 Backends
 from .almaden import FakeAlmaden
